Switch SIMD::Pointer::base from a Float* to Byte*
While having the base units as sizeof(float) is handy (as that's the smallest unit we currently support), it makes limit checking tricky as we're continually converting units of byte <-> float.

This sets us up nicely for when we want to support smaller types.

+++ src/Pipeline/SpirvShader.hpp
@@ -69,9 +69,9 @@ namespace sw
 			inline void addOffset(Int delta) { offset += delta; uniform = false; }
 
 			// Base address for the pointer, common across all lanes.
-			rr::Pointer<rr::Float> base;
+			rr::Pointer<rr::Byte> base;
 
-			// Per lane offsets from base.
+			// Per lane offsets from base in bytes.
 			// If uniform is false, all offsets are considered zero.
 			Int offset;
